# public/swagger.yaml
openapi: 3.0.0
info:
  title: 'Mente Livre API'
  description: 'API REST para el sistema de gestión de terapeutas y citas de Mente Livre'
  contact:
    name: 'Mente Livre'
    email: mentelivre.pe@gmail.com
  version: 1.0.0
servers:
  -
    url: 'https://backend.mentelivre.org/'
    description: 'Servidor de Desarrollo Local'
paths:
...
  '/therapists/{therapistId}/schedule-overrides':
    get:
      tags:
        - 'Schedule Overrides'
      summary: 'Obtener excepciones de horario de un terapeuta'
      operationId: getScheduleOverrides
      parameters:
        -
          name: therapistId
          in: path
          required: true
          schema:
            type: string
        -
          name: week_start_date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 'Lista de excepciones de horario'
    post:
      tags:
        - 'Schedule Overrides'
      summary: 'Crear excepción de horario'
      operationId: createScheduleOverride
      parameters:
        -
          name: therapistId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: 'Excepción creada exitosamente'
        '400':
          description: 'Datos inválidos'
      security:
        -
          bearerAuth: []
  '/therapists/{therapistId}/schedule-overrides/batch':
    post:
      tags:
        - 'Schedule Overrides'
      summary: 'Crear múltiples excepciones de horario'
      operationId: createScheduleOverridesBatch
      parameters:
        -
          name: therapistId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: 'Excepciones creadas exitosamente'
      security:
        -
          bearerAuth: []
  '/schedule-overrides/{id}':
    put:
      tags:
        - 'Schedule Overrides'
      summary: 'Actualizar excepción de horario'
      operationId: updateScheduleOverride
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Excepción actualizada exitosamente'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - 'Schedule Overrides'
      summary: 'Eliminar excepción de horario'
      operationId: deleteScheduleOverride
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Excepción eliminada exitosamente'
      security:
        -
          bearerAuth: []
  '/therapists/{therapistId}/schedule-overrides/week':
    delete:
      tags:
        - 'Schedule Overrides'
      summary: 'Eliminar todas las excepciones de una semana'
      operationId: deleteScheduleOverridesByWeek
      parameters:
        -
          name: therapistId
          in: path
          required: true
          schema:
            type: string
        -
          name: week_start_date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 'Excepciones eliminadas exitosamente'
      security:
        -
          bearerAuth: []
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Token JWT de autenticación. Incluir en el header: Authorization: Bearer {token}'
      bearerFormat: JWT
      scheme: bearer
...